Modified the default value for domain allowance.
Now the policy for the client is to let the code try to connect everywhere.
If somebody using the client in his app wants to limit the allowed/denied
domains has simply to overwrite the NAME_REDIRDOMAINDENY_RE,
NAME_CONNECTDOMAINALLOW_RE, NAME_CONNECTDOMAINDENY_RE,
NAME_REDIRDOMAINALLOW_RE in the client's XrdOucEnv before attempting
to connect. Through a call to XrdClient::SetParm.