[XrdCl] Implement fork safety
The cliet API classes are now fork-safe provided that the "RunForkHandler"
environment property is set to 1. Fork may occur at any time, even while
there are requests in the fly and the File and FileSystem objects remain
valid both in the parent and in the child. In the parent all the requests
are continued after the fork. In the child all in-the-fly requests are
cancelled (ie. their call-backs will not be called) and their heap memory
leaked. If there is no activity on-going when the fork happens, there is
no memory leak.