XrdCl: In recovery mode remove the kXR_delete flag but also add kXR_open_updt
When recovering a write error we need to remove the kXR_delete flag but we
also need to add the kXR_open_updt flag. If the user specified only the delete
flag when issuing the open then this will become an open for read in the back
without the client having any knowledge of this. When the user tries to write
he will get an error: bad file descriptor as he is trying to write to a file
opened for write. Partially addresses #287.