XrdCl: Append opaque info in case we retry at a data server after being redirected.
In case we were redirected to a data server and a retry is was performed
for a write operation the opaque info was missing from the try request.
Besides this, we now remove the kXR_delete and kXR_new flags only if
the first open on the data server was actually successful. Fixes #287.