Fixed race condition in multistream access to files fo CEPH
This was leading to file corruption and was mainly due to the absence of pread/pwrite interfaces in the layer in front of ceph. A non atomic combination of seek + read/write was used, leading to the problem.
My thanks to RAL for identify the problem and making most of the debugging.