[XrdCeph] Implemented pools of ceph objects
Allowed to use mutiple ceph objects in parallel when talking to ceph to workaround the limitation that a single threas is dealing with async answers per object.
In order to use this features, one has to add a line to the xrootd config file :
   ceph.nbconnections <n>
By default, only one object will be used.