XrdCl: Add option to load a client plugin for any url through the configuration specified in /etc/xrootd/client.plugin.d/.
This can be achieved by using "*" for the url key in the definition of the
client plugin. For example:
// file: /etc/xrootd/client.plugin.g/apply_to_all.conf
url = *
lib = /path/to/library
enabled = true

The same behaviour can also be obtained by using the XRD_PLUGIN env
variable. Using the env variable approach always has the highest
priority.