xrootdfs: remove default fuse argument "allow_other" as it is impossible to unset
Motivation:
This prevented mounting if user_allow_other is not set in fuse.conf
So a user with no root access could not use xrootdfs and could not change this argument without recompiling.
Change:
Don't pass allow_other to fuse by default.
Result:
Now xrootdfs works if user_allow_other is disabled, and allow_other argument can be passed when mounting if it is needed.