Make generated RPMs and version numbers monotonically increasing.
Working with RPMs and pre-releases is currently a pain - the output
of `makesrpm.sh` is not generally usable.  This is because the RPM
has distinct versioning schemes (4.6.1234 vs 20161215) depending
whether you are on an exact tag or not.  This means that upgrades
from a pre-release version to an official release is generally
impossible (at least, one wouldn't give a tester a pre-release out of
fear of messing up their system).

This introduces a new scheme for pre-releases:
  $MAJ.$MIN.$BUG+1-pre

where maj / min / bug versioning is based on the most recent tag.

For example, if the latest release is 4.6.1, then the pre-release
would be 4.6.2.pre.  If `makesrpm.sh` is run, then the RPM NVR would
be

4.6.2.pre-0.164.git.gbddc68e

where `164` is the number of commits since the 4.6.1 tag and gbddc68e
is the current git short hash.

NOTE: when master is pointing at a new release series, we should then
tag something along the lines of the next release.  For example, once
master is pointing at 4.7.x, we should make a tag of 4.6.99.

With this scheme, testers can cleanly upgrade from a pre-release to
a final release, reducing the likelihood they need to trash their sytem
setup in order to do testing.