[XrdCl] Avoid FileStateHandler deadlock while forking.
The recursive lock cannot be unlocked in the child atfork
handler, as the LWP id is used to identify the thread owner
in the current mutex implementation. LWP is unique per
system, and hence a thread in the child process cannot have
the same LWP as the thread locking the mutex in the parent
prepare atfork handler.