Another attempt at getting HTTP PUT flags right.
Originally, HTTP PUT could create but not overwrite.  While fixing
this, the flags were changed -- but it could overwrite and not create!

With this change, I have tested that both overwriting and creating
work.

We likely need additional follow-up because, with this change, a
user cannot perform HTTP PUT with the "create" authorization alone.