Refactor file interaction out of the state object.
Currently, there is a one-to-one correspondence between a HTTP request
(a state object) and a stream going out to a Xrootd file handle.

In preparation for multiple streams, we want to split up the concept
of a request and a file stream.  Multiple streams will be implemented
by doing multiple large requests going to the same file.