Close file handle for simple HTTP reads.
If a client requested a single byte range which was less than
the filesize, then the file handle was never explicitly closed.

When the TCP connection was closed, all open file handles on the
bridge are automatically closed too.  Hence, it seems like this is
a moot point - a transient "leak".  However, we have observed some
clients (particularly: CVMFS) that hold open TCP connections for
hours or days, resulting in the "leak" of a file handle per read
operation.  In some cases, the number of file handles leaked was in
the tens-of-thousands.

In addition to the obvious memory issues, this was observable in the
f-stream monitoring as the near-simultaneous closing of 10k file handles
caused UDP buffer overflows and dropped monitoring packets.