Revert "Merge pull request #679 from bbockelm/xrootd-tpc-rename"
This reverts commit e3550ce0af5ee08deae3d4ea660f4f1ed2d52ced, reversing
changes made to a768fd61f42d557bf80129eff88202fbe526c7a4.

It has been decided that this PR is to be reverted as it does not follow
XRootD coding conventions and brakes the build on vanilla RHEL6 platform
in presence of libcurl-devel package. There should be a check in the
CMake infrastructures that ensures the XrdTpc package will be only built
if a proper version of libcurl-devel is present (or XrdTpc should built
with version of libcurl-devel provided in raw RHEL6 environment).
Moreover the new component should be packaged properly (if neccessary
a _with_tpchttp macro should be tested in the spec file).