Use hostname, not reverse DNS, for address comparison.
This changes XrdSecgsi to prefer to use the hostname for the purpose
of matching a certificate to a hostname (as opposed to the prior
behavior of a reverse DNS lookup).

Relying on reverse DNS is considered insecure; note that all the
other security mechanisms use the hostname.

With the SAN changes allowing multiple potential patterns in the
certificate, admins should be able to handle all the potential use
cases.