Simplify logic for utilizing DNS.
Rely more on XrdNetAddr routines where at all possible.

We now call a hostname non-qualified if it contains no '.' characters.
While the previous algorithm potentially handled more side cases,
it had the strong downside of always relying on DNS security.  Since
that's precisely what we want to avoid, we only consider the case
where the user specifies `foo` and wants the search name to expand
it to `foo.example.com`.