Reset filesize when the XrdHttpReq object is used.
The `filesize` was not being reset between requests.  If it is non-zero,
the `XrdHttpReq` object uses the size when parsing header to potentially
reduce the total read size (preventing it from going past the EOF).

Unfortunately, since the `filesize` was set to the size of the *previous*
file transfer, this could cause the total response size to be less than
the request, resulting in a shorter-than-expected response to the client.

Backport of #814