Fallback to reverse DNS in all cases.
Whenever we have enabled trust DNS, we should fall back to using
reverse DNS in all cases, not just when an IP address is given.

This provides backward compatibility with the case where the client
uses a DNS alias directly but the server is missing the SAN for
that DNS alias (instead the server only has the canonical name
for the IP address in the DN).