Add standalone pyxrootd setup script
This is useful in the case where there is already an xrootd install but
it does not include the python portion.  This happens e.g. if xrootd is
installed in a python2 context and the user wants also to use it for
python3.  One can simply download the source and run in this directory:
> python setup_standalone.py install