Allow the state and stream objects to explicitly finalize.
This ensures the user of the object is aware when the underlying file
has closed.

When the file is closed in the stream's destructor, we may have already
sent the response back to the client.  For filesystems that might
buffer significant amounts of data in the client library, it's possible
that subsequent operations on a different server attached to the same
distributed file system may see incorrect file sizes (as the open file
still hasn't been closed successfully).