Allow admin to customize default macaroons authz.
With this, the admin can set a new `macaroons.onmissing` configuration
flag to allow the authorization plugin to always allow in the case
when no macaroon exists.

This is useful for EOS as EOS does not utilize the XRootD authorization
system and will apply its own authorizations downstream of this plugin.