Always query dual stack for HTTP TPC.
If we open a file using the SFS interface, the OFS plugin will, by default,
only query for servers that support IPv6.  If a cluster only has IPv4
addresses, then this will always fail.

This changes the default to dual-stack: the transfer can be serviced by
a server that speaks either IPv4 or IPv6 (or both!).  This is likely the
best we can do as we have no indication of whether the source side is IPv4,
IPv6, or dual stack itself.

Fixes #968