Fix two problems with RFC-3230 handling in xrootd:
First, the Digest response header copies the supplied Want-Digest
request header value.  A request with:

    Want-Digest: MD5,ADLER32

would yield a response like:

    Digest: MD5,ADLER32=<md5-value>

where the correct response would be:

    Digest: MD5=<md5-value>

Second, the method that calculates whether the response should be
encoded in hexadecimal or base64 expects an algorithm name, but is given
the Want-Digest request header value.  If the client makes a request with:

    Want-Digest: MD5, ADLER32

then the response is incorrectly encoded using hexadecimal.