Use an FHS 3.0 compliant pidpath
https://en.wikipedia.org/wiki/Filesystem_Hierarchy_Standard#FHS_compliance

> Modern Linux distributions include a /run directory as a temporary
filesystem (tmpfs) which stores volatile runtime data, following the
FHS version 3.0. According to the FHS version 2.3, this data should be
stored in /var/run but this was a problem in some cases because this
directory isn't always available at early boot. As a result, these
programs have had to resort to trickery, such as using /dev/.udev,
/dev/.mdadm, /dev/.systemd or /dev/.mount directories, even though the
device directory isn't intended for such data. Among other advantages,
this makes the system easier to use normally with the root filesystem
mounted read-only.