Add support for buffering even when single-stream TPC
Currently, we take a write from libcurl and immediately push it
down to the underlying filesystem layer.  libcurl will issue fairly
small writes, especially if the underlying filesystem is actually
a remote filesystem (meaning the small write may be performed
synchronously instead of just going to the OS page cache).

This code aggregates writes in memory until there's a full 1MB
available; the MB is then pushed to the filesystem layer.

Note: as a side-effect of this change, writes become aligned on
MB boundaries.  This avoids an issue with xrootd-ceph which does
not currently support unaligned writes for EC setups.