[XrdNet] Correctly use IN6_IS_ADDR_V4MAPPED macro.
In case of `ai_family == AF_INET6` the `ai_addr` points to a memory block containing a `sockaddr_in` struct.
However, the `IN6_IS_ADDR_V4MAPPED` macro expects `in6_addr` struct, hence we need to pass `sockaddr_in::sin6_addr`.

This was compiling well on Linux platforms because in the Linux implementation the pointer is being explicitly cast:
```
  (__extension__                                                              \
   ({ const struct in6_addr *__a = (const struct in6_addr *) (a);             \
      __a->s6_addr32[0] == 0                                                  \
      && __a->s6_addr32[1] == 0                                               \
      && __a->s6_addr32[2] == htonl (0xffff); }))
```

However, on OSX platform the builds were failing as the definition is slighly diffrent:
```
	((*(const __uint32_t *)(const void *)(&(a)->s6_addr[0]) == 0) && \
	(*(const __uint32_t *)(const void *)(&(a)->s6_addr[4]) == 0) && \
	(*(const __uint32_t *)(const void *)(&(a)->s6_addr[8]) == \
	ntohl(0x0000ffff)))
```