Use correct size for struct
The fix for the gcc 11 compiler warning in commit
8f36753d46c183aa89058219e36e9a2980afd250 is incorrect.
Placing the sizeof statement in parenthesis does not correct the bug,
it just obfuscates the code so that the compiler doesn't see it.

The correct fix is to give the unnamed struct a name so that the
sizeof the struct can be used in the calculation.

Conflicts:
	src/XrdOuc/XrdOucPsx.cc