XrdTpc: Factor out recursive writes.
This separates the writing methods of the stream to avoid potential
recursion when dumping buffered data.  This also allows the buffer
to accept partial data so the beginning offset of the buffer is
always a multiple of the buffer size (i.e., the buffer offsets
always stay MB-aligned).

With this, a single incoming buffer from curl may be split across
two entries in order to keep alignment.