Add a crude timeout mechanism.
This does not enforce a minimum bandwidth (N MB/s) but rather
aborts the transfer if zero data has moved within a given timeframe.
By default, the transfer has 120 seconds to move the first byte of
data and it must transfer at least 1 byte every 60 seconds thereafter.

This purposely does not do bandwidth minimums as that's built into
future versions of `curl` and I'd prefer to not re-implement it.

The default timeouts can be overridden by setting:

```
tpc.timeout timeout_val [first_byte_timeout_val]
```

where the default is equivalent to:

```
tpc.timeout 60
```

If the `first_byte_timeout_val` is not set, it defaults to 2x the
`timeout_val`.