XrdTls: Refactor temp CA code to use ADMINPATH directory.
This overhauls the temp CA code to use the XRDADMINPATH directory;
since we can assume XRootD "owns" the directory, we don't need to use
a temporary filename for the CA.  Thus, we don't need to generate secure
filenames in /tmp and there's no longer a need to track CA file lifetime.
This simplifies the shared pointer to simply a coordination with the
maintenance thread; libcurl can reuse the same filename for as long as
possible.