Fix  AIO read logic affecting end of file
Fix the logic when doing AIO reads. rc returns the number of bytes read; which would be 0 for end of file, or, less than nbytes if a partial read was required. 
Previously, the number of bytes actually read was always returned, except if 0 bytes were read, in which case the requested number of bytes was returned. 
Now, the number of bytes actually read (including 0) is always returned into the callback. 
There remains a final 0 byte read, but this should stop excess data being returned.