Do not hardcode error numbers
The error numbers defined on Linux can not be reused on other systems
since this causes clashes.

On Linux ECHRNG is 44, but defining ECHRNG to be 44 on systems that do
not define ECHRNG is not appropriate. On e.g. kFreeBSD this will clash
with ESOCKTNOSUPPORT, which is defined as 44 on that system.