Add max file open tracking.
This tracks the maximum number of files open per "entity" (which might
be distinct from the login name in case of tokens); once the file limit
is hit, the code returns an error instead of opening a new file.

Code is only active when throttle module is loaded and max is explicitly
set.  Here's an example configuration with a ridiculously low max (which
is useful for testing, of course):

```
throttle.max_open_files 2
xrootd.fslib ++ throttle
throttle.trace debug files
```