Allow explicit grid-mapfile entries to override voms-mapfile.
If there's an entry found in the corresponding grid-mapfile, then
that is considered "more specific" and overrides the "less specific"
voms-mapfile entry.  E.g., a mapping to `bbockelm` should be picked
over a generic `cms`.

To differentiate between an explicit mapping and the fallbacks
generated for the plugins, we use the entity's extended attribute
mechanism.