[CI] Move xrootd.egg-info off PYTHONPATH
As discovered in XRootD Issue #1604 there are instances where if a Git tag
is created, then the current xrootd.egg-info that is generated in the top
level of the repository by the current packaging/wheel/publish.sh workflow
will pollute the PYTHONPATH (which includes the current working directory by
default --- this is one of the reasons src directory structures are currently
recommended by the PyPA) causing problems with trying to install the built sdist
from the top level of the repo with `python3 -m pip install dist/xrootd*.tar.gz`.

Ideally the publish.sh workflow should be refactored to make this not a problem,
but for the time being just use the easy workaround of changing paths so that the
xrootd.egg-info directory is not on PYTHONPATH and install the sdist normally.