Make the `acc` subsystem aware of request-based name mapping.
When a per-request mapping is made (such as with SciTokens), the
mapped name is not put in the typical `Entity.name` but rather an
extended attribute.  Only authorization plugins that have been
explicitly updated to look for this attribute can be usefully chained
after per-request mapping.

These changes allow the `acc` subsystem to be chained in a useful
manner after a token plugin (such as SciTokens).