Populate XrdSciTokens with more detailed log messages.
This introduces configuration-file-controlled tracing of SciTokens-
related messages and a number of detailed log messages.

As typical, this is controlled by a new config variable:
```
scitokens.trace [level] [level] ...
```

Where level can be one of `all`, `none`, `debug`, `info`, `warning`, or
`error`.

Roughly, the levels imply:
- Error: fatal issue in plugin configuration
- Warning: Potentially-problematic configuration or token that is invalid
- Info: Informational messages about the plugin.  Shouldn't have any logging
  per-request
- Debug: Verbose, per-request details about the tokens.

The default is error + warning; debug is not recommended for continuous
production settings but is immensely helpful in solving problems.