[cmake] Add PIP_OPTIONS option for Python bindings install control
Add CMake option PIP_OPTIONS to allow the user to pass in pip options
at build configuration time to customize the Python bindings install.
As '--verbose' is a pip option, this makes the PIP_VERBOSE CMake option
added in PR 1586 redundant, and so it is removed. Additionally, as '--prefix'
is also a pip option, append the selected default value of
'$ENV{DESTDIR}/{CMAKE_INSTALL_PREFIX}' as a '--prefix' pip option if there is
no '--prefix' given by the user. If there is a given '--prefix' from the user in
PIP_OPTIONS already, warn the user that this might not be a good idea unless
they know what they're doing.

Adrian provided the idea of providing the user a PIP_OPTIONS CMake variable at build
configure time.
Henry provided help on getting PIP_OPTIONS to be recognized as a string by defult.

Co-authored-by: Adrian Sevcenco <adrian.sevcenco@cern.ch>
Co-authored-by: Henry Schreiner <henry.fredrick.schreiner@cern.ch>