[packaging] Use 'command -v' over 'which' for portability
'command -v' is a POSIX safe way to find an executable program.
While 'which' is very common, it still requires installation on many
Linux operating systems, and so 'command -v' can offer a more portable
alternative with a very low amount of additional cognitive overhead on
the maintainers.

Additionally, use a subshell, $(), over backticks for command substitution
as $() is recommended for multiple reasons including readability.
c.f.:
* https://stackoverflow.com/a/4708569/8931942
* Why is $(...) preferred over `...` (backticks)?
   - http://mywiki.wooledge.org/BashFAQ/082
* POSIX spec for Command Substitution
   - https://pubs.opengroup.org/onlinepubs/9699919799/xrat/V4_xcu_chap02.html#tag_23_02_06_03