[FIX] Use python's sys.version_info.minor to ensure correct library
To avoid finding multiple directories that have 'python' in the name
under /usr/local/lib/ on the GitHub Actions runners, use Python's
sys.version_info.minor to get the minor version of python3. As we know
that we're using Python 3, the minor version is the only piece of information
that needs to be determined.