[pip] Remove manual renaming of egg-info to dist-info
As all CPython versions that have 'shutil.which' (Python 3.3+) are able to run
packaging/wheel/setup.py then they will be able to take advantage of having pip
be able to create a valid wheel during install time. In the event that a very old
version of pip is used, an egg distribution will be installed but it will still work
fine so there is no real reason to rename it in a manner that mocks the dist-info of a
wheel.