Introduce 'exclusive' operations.
With this, file creation and rename can be performed with a lower-
privileged 'exclusive' authorization level.

Code paths are unchanged for existing authorization but, on authz
failure, we will test for these new lower-privileges and, if present,
allow the creation to continue provided there is no preexisting data
that will be overwritten.