Compilable first draft of Repair Tool
Almost working except for the repair tool being destroyed due to a shared pointer

This version works quite well, but uses mtfile

Working version of the repair tool handles data and metadata corruption
rebased to master

Cleaned code
rebased to master

Last fixes
rebased to master

Restructured and commented repair tool functionality, the opening of archives can either be followed by a host replacement or on its own to detect and report on damaged archives to the user

Added callback for just checking a file's checksum, but has a bug with checksum comparison in reader

Fixed the archive opening and closing and the only instance that checks metadata is now the repair tool

Archive closing waits until everything is written using a pipeline final and condition locking

Fixed: We wait until all writes are finished before closing archives, WriteIntoFile is now a zip operation

Fix: Wait for all replacement archives to be opened

Added a test that corrupts two archives (or one archives in two spots), ran it successfully a couple hundred times

Fixes: Corrupt archives at correct offsets and replaced two usages of uncompressedSize by compressedSize

Fixed the actual root of the issue because corrupted offsets can lead to signflips in minus operations

Small concurrency fix and formatting

Concurrent version of repair

Fixes to make more than nbdata+nbparity hosts possible, changed unit test to use 2 hosts more

Cleaned some code

Further cleanup

Removed more unnecessary code and commented out lines

Added a test to confirm that repair fails when too much is corrupted

Added timeouts to the repair tool and a check in the opening process that checks whether one of the archives was blocked by the user. If it is the case, the tool exits immediately

Adding some timeouts seems to introduce a bug, might have to revert?

Trying to fix segfault for missing host

Fixed an error that the open pipeline would fail if not all readHealth are successful

Fixed the bugs, forgot a handler callback

Changed atomics to relaxed memory order and implemented a block pool to limit the memory consumption

Corrected a rebase mistake

Removed file repair functionality to split it into a second PR