[CI] Perform a local installation of Python bindings for testing on Mac
This is required as by default the installed Python doesn't know about
XRootD which was installed into /usr/local/lib/python3.X. Using a local
installation we can work around it. Note that we cannot set --user in
PIP_OPTIONS as it conflicts with --prefix, which is added by the build
system in bindings/python/CMakeLists.txt.