[CMake] Add check for musl libc
Out of principle musl libc provides no means for detecting it, leaving
developers to figure it out for themselves. The choice here is to use
the compiler to print the target triple, which contains the libc
implementation as the last part. On Voidlinux it's x86_64-linux-musl,
and on Alpine it's x86_64-alpine-linux-musl. This should be reliable
enough to not wrongly detect other libc implementations as musl (and
failure to detect will only affect platforms based on musl as well).

See https://wiki.musl-libc.org/faq.html for more information.