Add an error object to XrdOss.
While XrdOss has a simple POSIX-like interface, in reality the
underlying implementation can be quite complicated -- XrdPfc, for
example.  Accordingly, we should allow for a richer error message
interface between the OSS and its callers.  This adds a new virtual
function which the SFS interface can invoke to get additional error
information.

The intent is this will be used by XrdPfc and allow clients to
differentiate between "the cache gave you a permission denied" and
"the origin gave the cache a permission denied".

*A note on ABI compatibility*: Technically, this is an ABI change
and hence would need to wait until XRootD 6 to be deployed.  However,
this takes advantage of a quirk of GCC: the ordering of the virtual
functions are preserved.  Hence, the new functions are placed out-of-order.